+++ packages/fuselage-hooks/jest.config.js
@@ -1,3 +1,7 @@
 module.exports = {
   preset: 'ts-jest',
+  errorOnDeprecated: true,
+  testMatch: [
+    '**/src/**/*.spec.[jt]s?(x)',
+  ],
 };
