+++ packages/fuselage-hooks/src/useDebouncedValue.spec.ts
@@ -1,11 +1,10 @@
 import { useState } from 'react';
 
-import { runHooks } from '../jestHelpers';
-
-import { useDebouncedValue } from '..';
+import { runHooks } from './jestHelpers';
+import { useDebouncedValue } from '.';
 
 describe('useDebouncedValue hook', () => {
-  let delay;
+  let delay: number;
   beforeEach(() => {
     jest.useFakeTimers();
     delay = Math.round(100 * Math.random());
