+++ packages/fuselage-hooks/src/useLazyRef.spec.ts
@@ -1,6 +1,5 @@
-import { runHooks } from '../jestHelpers';
-
-import { useLazyRef } from '..';
+import { runHooks } from './jestHelpers';
+import { useLazyRef } from '.';
 
 describe('useLazyRef hook', () => {
   it('returns the computed value immediately', () => {
