+++ packages/fuselage-hooks/src/useMutableCallback.server.spec.ts
@@ -2,9 +2,8 @@
  * @jest-environment node
  */
 
-import { runHooksOnServer } from '../jestHelpers';
-
-import { useMutableCallback } from '..';
+import { runHooksOnServer } from './jestHelpers';
+import { useMutableCallback } from '.';
 
 describe('useMutableCallback hook on server', () => {
   it('returns a callback that invokes the mutable one', () => {
