+++ packages/fuselage-hooks/src/useResizeObserver.server.spec.ts
@@ -2,9 +2,8 @@
  * @jest-environment node
  */
 
-import { runHooksOnServer } from '../jestHelpers';
-
-import { useResizeObserver } from '..';
+import { runHooksOnServer } from './jestHelpers';
+import { useResizeObserver } from '.';
 
 describe('useResizeObserver hook on server', () => {
   it('immediately returns undefined sizes', () => {
