+++ packages/fuselage-hooks/src/useSafely.spec.ts
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 
-import { runHooks } from '../jestHelpers';
-
-import { useSafely } from '..';
+import { runHooks } from './jestHelpers';
+import { useSafely } from '.';
 
 describe('useSafely hook', () => {
   it('returns a new updater that invokes the previous one', () => {
