+++ packages/fuselage-hooks/src/useToggle.spec.ts
@@ -1,6 +1,5 @@
-import { runHooks } from '../jestHelpers';
-
-import { useToggle } from '..';
+import { runHooks } from './jestHelpers';
+import { useToggle } from '.';
 
 describe('useToggle hook', () => {
   it('has false value when an initial value is undefined', () => {
